refactor: extract shared tenant resolution helpers into server/tenant/resolve.ts
Deduplicate resolveAgentTenant (init.ts:231-235) and resolveAgentTenant
(broadcasting.ts:36-41) into a single shared utility. Also extracts
resolveCouncilTenant. The new function uses DEFAULT_TENANT_ID constant
and accepts an optional multiTenant parameter (defaults to true).

Resolves the TODO at init.ts:230.
